chore: extract block form loading to a composable
